refactor(package): changed args package to config to reflect its scope
