+++ nri-config-generator/config/args_test.go
@@ -1,4 +1,4 @@
-package args
+package config
 
 import (
 	"io/fs"
