+++ nri-config-generator/config/loader.go
@@ -1,4 +1,4 @@
-package args
+package config
 
 import (
 	"os"
